@@ -8,25 +8,24 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.4|^8.0",
-        "ccxt/ccxt": "^1.53",
-        "fideloper/proxy": "^4.4",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "laravel/fortify": "^1.7",
-        "laravel/framework": "^8.40",
-        "laravel/sanctum": "^2.11",
-        "laravel/tinker": "^2.5"
+        "php": "^8.0.2",
+        "ccxt/ccxt": "^1.72.55",
+        "fruitcake/laravel-cors": "^v2.0.5",
+        "guzzlehttp/guzzle": "^7.4.1",
+        "laravel/fortify": "^v1.10.1",
+        "laravel/framework": "^v9.0.0",
+        "laravel/sanctum": "^2.14.0",
+        "laravel/tinker": "2.7.0"
     },
     "require-dev": {
-        "barryvdh/laravel-ide-helper": "^2.10",
-        "facade/ignition": "^2.5",
-        "fakerphp/faker": "^1.9.1",
-        "laravel/sail": "^1.0.1",
-        "mockery/mockery": "^1.4.2",
-        "nunomaduro/collision": "^5.0",
-        "nunomaduro/larastan": "^1.0",
-        "phpunit/phpunit": "^9.3.3"
+        "barryvdh/laravel-ide-helper": "^v2.12.2",
+        "spatie/laravel-ignition": "^1.0.3",
+        "fakerphp/faker": "^v1.19.0",
+        "laravel/sail": "^v1.13.2",
+        "mockery/mockery": "^1.5.0",
+        "nunomaduro/collision": "^v6.1.0",
+        "nunomaduro/larastan": "^2.0.0",
+        "phpunit/phpunit": "^9.5.13"
     },
     "autoload": {
         "psr-4": {
